Update NOTAM model to use Double for durationHours and adjust related tests
- Changed the type of `durationHours` in the `NOTAMSchedule` model from `Int?` to `Double?` for more precise duration representation.
- Updated test cases in `NOTAMParsingTests` to reflect the new duration format, ensuring accurate decoding and validation of NOTAM data.

+++ Buzz/Models/NOTAM.swift
@@ -196,7 +196,7 @@ struct NOTAMSchedule: Codable, Identifiable {
     var id: String { source }
     let description: String
     let source: String
-    let durationHours: Int?
+    let durationHours: Double?
     let isSunriseSunset: Bool
     
     enum CodingKeys: String, CodingKey {
